refactor(core): MCP/Skill 从 core 完整迁出（Phase 1-4）#199
按方案推进 Phase 1-4，core 最终不持有 MCP/Skill 领域概念。

## Phase 1：清理无风险残留
- runtime.rs: MCP shim 过期注释改为 mcp__server__tool 显式分发说明
- core/mod.rs: 工具冲突注释改为通用插件冲突说明（不再特指 MCP）
- model.rs: 删除 core system prompt 的 Skill 规则条目（迁到 skill plugin prompt）
- AppServices: 删除 skill_registry 字段（无读取方，SkillPlugin 自托管）
- lifecycle.rs: 删除 skill_registry 构造

## Phase 2：AppPaths 清理
- 删除 skills_config_path / mcp_config_path / mcp_capability_cache_path 字段
  （无读取方，mcp/skill plugin 各自有 paths 模块）
- utils.rs 删除对应 default_*_path 函数
- lifecycle.rs load_or_default 不再构造这三个路径；首次安装检测改为只查 app.json

## Phase 3：mcp-lock 迁出 core
- 新建 tiangong-plugin-skill/src/mcp_lock.rs：sync_mcp_dependency_lock 逻辑
  + McpDependencyLockRecord 类型 + default_mcp_lock_path（迁自 core locks.rs）
- SkillPlugin 加 sync_mcp_dependency_lock() 固有方法
- core 删除 app_state/repository/locks.rs + AppRepository::sync_mcp_dependency_lock
- core 删除 support.rs 的 McpDependencyLockRecord + utils.rs 的 default_mcp_lock_path
- 删除 storage.rs 的 persist_agent_configs_only / persist_agent_configs_no_merge_mcp
  （已退化为 mcp-lock 代理，无存在必要）
- lifecycle.rs 删除 sync_mcp_dependency_lock 调用（改由入口层在 skill 管理时调 plugin）

## Phase 4：Skill 领域模块迁出 core
- core/src/skill/*（registry/package/analysis/context/init/util）→ plugin-skill/src/skill_*.rs
- agent_config.rs 的 Skill 类型（SkillsConfig/InstalledSkillConfig/SkillSourceConfig/
  SkillMcpRequirementConfig/SkillPermissionConfig）→ plugin-skill/src/skill_config.rs
- core 删除 pub mod skill + pub mod agents（agents/skill_convert_agent.rs 是死代码，一并删）
- agent_config.rs 精简为仅 AgentConfig
- plugin-skill lib.rs re-export Skill 领域类型
- 下游 crate 调整：
  - tiangong-config: SkillsConfig 改从 tiangong_plugin_skill 引用（+依赖）
  - tiangong-entry: skill 类型改从 plugin-skill 引用
  - src-tauri/view.rs: LoadedSkill/InstalledSkillConfig 改从 plugin-skill 引用
- plugin-skill Cargo.toml: edition 2024 + 加 toml/scru128/serde/tracing 依赖

验收：cargo check --workspace + cargo clippy --workspace 零 error 零 warning；
core/plugin-skill/plugin-mcp/config/entry/cli/server 测试全部通过
（skill plugin 新增 11 个 registry 测试）。

core 搜索 mcp|skill 现仅剩通用注释/路径函数名/权限测试字符串。
