+++ crates/plugins/tiangong-plugin-skill/Cargo.toml
@@ -1,15 +1,19 @@
 [package]
 name = "tiangong-plugin-skill"
 version.workspace = true
-edition = "2021"
+edition = "2024"
 license = "Apache-2.0"
 
 [dependencies]
 tiangong-core = { workspace = true }
 tiangong-types = { workspace = true }
 anyhow = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tokio = { workspace = true }
+tracing = { workspace = true }
+toml = { workspace = true }
+scru128 = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
