+++ crates/plugins/tiangong-plugin-skill/src/skill_analysis.rs
