+++ crates/plugins/tiangong-plugin-skill/src/skill_init.rs
