+++ crates/plugins/tiangong-plugin-skill/src/skill_util.rs
