+++ crates/tiangong-core/src/app_state/support.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use super::*;
 
 #[derive(Debug, Serialize, Deserialize, Default)]
@@ -35,22 +33,9 @@ pub(in crate::app_state) struct LoadedState {
 
 pub use tiangong_types::StreamEvent;
 
-/// mcp-lock.json 中单个 MCP 依赖包的引用计数记录。
-///
-/// 语义：聚合所有已安装 skill 的 `requires.mcp` 声明，按 `package[@version]` 分组，
-/// 记录被多少 skill 引用。不再记录安装路径与时间（旧字段的兼容读取见 serde default）。
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub(in crate::app_state) struct McpDependencyLockRecord {
-    /// 引用该依赖的 skill 数量。
-    pub(in crate::app_state) ref_count: usize,
-}
-
 #[derive(Debug)]
 pub struct AppPaths {
     pub app_storage_path: PathBuf,
-    pub skills_config_path: PathBuf,
-    pub mcp_config_path: PathBuf,
-    pub mcp_capability_cache_path: PathBuf,
     pub sessions_dir_path: PathBuf,
 }
 
@@ -59,5 +44,4 @@ pub struct AppServices {
     pub repository: AppRepository,
     pub runtime: RuntimeEngine,
     pub turn_service: AppTurnService,
-    pub skill_registry: Arc<crate::skill::SkillRegistry>,
 }
