+++ crates/tiangong-core/src/app_state/tests/repository.rs
@@ -178,54 +178,3 @@ fn load_from_disk_filters_child_sessions_from_session_list_state() -> Result<()>
         },
     )
 }
-
-#[test]
-#[ignore = "Phase 4: mcp-lock 逻辑迁移到 plugin 后重写"]
-fn sync_mcp_dependency_lock_writes_expected_ref_counts() -> Result<()> {
-    with_isolated_state("tiangong-mcp-dependency-lock", |paths, state| {
-        // skills 已脱离 agent_config，sync_mcp_dependency_lock 从磁盘 registry 扫描。
-        // 在隔离存储目录下创建两个 skill 的 skill.toml，声明 mcp 依赖。
-        let skills_dir = paths.fake_home.join(".tiangong").join("skills");
-        fs::create_dir_all(skills_dir.join("alpha"))?;
-        fs::create_dir_all(skills_dir.join("beta"))?;
-        fs::write(
-            skills_dir.join("alpha").join("skill.toml"),
-            "id = \"alpha\"\nname = \"Alpha\"\nversion = \"0.1.0\"\nentry = \"SKILL.md\"\n\n[source]\ntype = \"local\"\nvalue = \"\"\n\n[[requires.mcp]]\nid = \"m1\"\nsource = \"\"\npackage = \"pkg-a\"\nversion = \"1.0.0\"\n\n[[requires.mcp]]\nid = \"m2\"\nsource = \"\"\npackage = \"pkg-b\"\nversion = \"\"\n",
-        )?;
-        fs::write(
-            skills_dir.join("beta").join("skill.toml"),
-            "id = \"beta\"\nname = \"Beta\"\nversion = \"0.1.0\"\nentry = \"SKILL.md\"\n\n[source]\ntype = \"local\"\nvalue = \"\"\n\n[[requires.mcp]]\nid = \"m3\"\nsource = \"\"\npackage = \"pkg-a\"\nversion = \"1.0.0\"\n",
-        )?;
-
-        state.sync_mcp_dependency_lock()?;
-
-        let mcp_lock_path = skills_dir.join("mcp-lock.json");
-
-        let mcp_lock: std::collections::BTreeMap<String, serde_json::Value> =
-            serde_json::from_str(&fs::read_to_string(mcp_lock_path)?)?;
-
-        assert!(
-            !paths
-                .fake_home
-                .join(".tiangong")
-                .join("skills")
-                .join("skills-lock.json")
-                .exists()
-        );
-        assert_eq!(
-            mcp_lock
-                .get("pkg-a@1.0.0")
-                .and_then(|item| item.get("ref_count"))
-                .and_then(|value| value.as_u64()),
-            Some(2)
-        );
-        assert_eq!(
-            mcp_lock
-                .get("pkg-b")
-                .and_then(|item| item.get("ref_count"))
-                .and_then(|value| value.as_u64()),
-            Some(1)
-        );
-        Ok(())
-    })
-}
