+++ crates/tiangong-core/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod agent_config;
 pub mod agent_input;
 pub mod agent_team;
-pub mod agents;
 pub mod app_state;
 pub mod approval_store;
 pub mod browser_trait;
@@ -26,7 +25,6 @@ pub mod react;
 pub mod runtime;
 pub mod runtime_env;
 pub mod session;
-pub mod skill;
 mod stream_throttle;
 pub mod task;
 pub mod terminal_trait;
