--- crates/tiangong-core/src/skill/mod.rs
@@ -1,18 +1,0 @@
-mod analysis;
-mod context;
-mod init;
-mod package;
-mod registry;
-mod util;
-
-pub use analysis::analyze_external_skill;
-pub use context::build_skill_hints;
-pub use init::init_tiangong_skill_scaffold;
-pub use package::{
-    SkillConversionArtifacts, load_skill_from_local_dir, prepare_skill_source_for_install,
-};
-pub use registry::{
-    DEFAULT_SKILL_REGISTRY_CACHE_TTL, DEFAULT_SKILL_REGISTRY_LOADED_CAPACITY, LoadedSkill,
-    SkillManifest, SkillRegistry, SkillRegistryEntry, SkillRegistryIssue, SkillRegistryIssueKind,
-    SkillRegistryView, read_skill_manifest, scan_skill_registry, write_skill_available,
-};
